js: converted team members files to vue format
