--- app/views/teams/components/members/_form.html.slim
@@ -1,33 +1,0 @@
-#new-team-member-form ref="form" class="collapse"
-  = form_for :team_user, url: team_users_path, html: { class: "form-horizontal", role: "form", "@submit.prevent" => "onSubmit", novalidate: true } do |f|
-    .form-group
-      = f.label :role, { class: "control-label col-md-2" }
-      .col-md-7
-        = f.select(:role, TeamUser.roles.keys.map { |role| [role.titleize,role] }, {}, {class: "form-control", "v-model" => "member.role"})
-    .form-group.has-feedback :class=="{ 'has-error': $v.member.user.$error }"
-      = f.label :user, { class: "control-label col-md-2" }
-      .col-md-7
-        vue-multiselect{
-          class="team_user_user"
-          :class=="{'multiselect--selected': selectedMember != null}"
-          v-model="selectedMember"
-          label="name"
-          track-by="name"
-          placeholder="Type to search"
-          :loading="isLoading"
-          :options="members"
-          :max-height="400"
-          @close="onTouch"
-          @select="onSelect"
-          @remove="onRemove"
-          @search-change="searchMember"
-        }
-          span slot="noResult"
-            | Oops! No username found. Consider changing the search query.
-        span.help-block
-          span v-if="!$v.member.user.required"
-            | User can't be blank
-
-    .form-group
-      .col-md-offset-2.col-md-7
-        = f.submit('Add', class: "btn btn-primary", ":disabled" => "$v.$invalid")
