--- app/views/teams/components/members/_panel.html.slim
@@ -1,19 +1,0 @@
-.panel.panel-default
-  .panel-heading
-    .row
-      .col-sm-6
-        h5
-          a[data-placement="right"
-            data-toggle="popover"
-            data-container=".panel-heading"
-            data-content="<p>There are three types of users inside of a team:<br><strong>Viewer:</strong> has read only access. Only pull operations are permitted.<br/><strong>Contributor:</strong> has read and write access. Both pull and push operations are permitted.<br/><strong>Owner:</strong> like 'contributor', but can also manage the team.</p>"
-            data-original-title="What's this?"
-            tabindex="0" data-html="true"]
-            i.fa.fa-info-circle
-          '  Members
-      .col-sm-6.text-right
-        <toggle-link text="Add members" :state="state" state-key="membersFormVisible" class="toggle-link-new-member"></toggle-link>
-
-  .panel-body
-    .table-responsive
-      <team-members-table :members="members" sort-by="id" :current-member="currentMember"></team-members-table>
