--- app/views/teams/components/members/_table.html.slim
@@ -1,23 +1,0 @@
-div
-  table.table.table-striped.table-hover :class=="{'table-sortable': sortable}"
-    colgroup v-if="canManage"
-      col.col-5
-      col.col-40
-      col.col-35
-      col.col-10
-      col.col-10
-    colgroup v-if="!canManage"
-      col.col-5
-      col.col-50
-      col.col-45
-    thead
-      tr
-        th
-        th User
-        th Role
-        th v-if="canManage" Edit
-        th v-if="canManage" Remove
-    tbody
-      <team-members-table-row v-for="member in filteredMembers" :key="member.id" :member="member" :can-manage="canManage"></team-members-table-row>
-
-  <table-pagination :total.sync="members.length" :current-page="currentPage" :itens-per-page.sync="limit" @update="updateCurrentPage"></table-pagination>
