+++ app/views/teams/index.html.slim
@@ -1,3 +1,1 @@
 <teams-index-page :teams-ref="#{@teams_serialized}" teams-path="#{teams_path}" :is-admin="#{current_user.admin?}" :can-create="#{can_create_team?}" :owners-ref="#{@owners_serialized}" :current-user-id="#{current_user.id}">
-
-= render "teams/partials/components_templates"
