--- app/views/teams/partials/_components_templates.html.slim
@@ -1,13 +1,0 @@
-- if action_name == 'show'
-  - content_for :js_body do
-    script#js-team-members-panel-tmpl type="text/x-template"
-      = render "teams/components/members/panel"
-
-    script#js-team-members-table-tmpl type="text/x-template"
-      = render "teams/components/members/table"
-
-    script#js-team-members-table-row-tmpl type="text/x-template"
-      = render "teams/components/members/table_row"
-
-    script#js-team-member-form-tmpl type="text/x-template"
-      = render "teams/components/members/form"
