+++ app/views/teams/show.html.slim
@@ -1,4 +1,3 @@
 <teams-show-page :team-ref="#{@team_serialized}" :namespaces-ref="#{@team_namespaces_serialized}" :members-ref="#{@team_users_serialized}" namespaces-path="#{namespaces_path}" :current-member-ref="#{@current_member_serialized}" :manage-teams-enabled="#{manage_teams_enabled?}" :available-roles="#{@available_roles}"></teams-show-page>
 
-= render 'teams/partials/components_templates'
 = render 'namespaces/partials/components_templates'
